+++ data.py
@@ -1,0 +1,12 @@
+Students = []
+
+Teachers = []
+
+subjects = {'matemática': [], 
+            'artes': [],
+            'ciência': [],
+            'Educação Física': [],
+            'português': [],
+            'inglês': [],
+            'geográfia': [],
+            'história': []}
